fixed colab

+++ mlp/fixed_colab_copy.py
@@ -546,7 +546,7 @@ def is_triplet_set(key_list, index):
 
 def categorize_triplet(sequence, index):
     # Extract last 3 elements of each card in the triplet
-    
+    sequence = sequence.astype(int)
     card1 = tuple(sequence[0:12][index*3: (index + 1) * 3])
     card2 = tuple(sequence[12:24][index*3: (index + 1) * 3])
     card3 = tuple(sequence[24:36][index*3: (index + 1) * 3])
